Folder restructured, code adapted accordingly
