--- allure-results/0a38b116-1cde-437a-9756-19ff868325b8-result.json
@@ -1,1 +1,0 @@
-{"uuid":"0a38b116-1cde-437a-9756-19ff868325b8","name":"Test3 login to page @regression","historyId":"9ba47862b673822ab6987e4d6c14df9f:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861453335,"name":"browserType.launch","stop":1739861453495}],"attachments":[],"parameters":[],"start":1739861453329,"name":"fixture: browser","stop":1739861453495},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861453503,"name":"browser.newPage","stop":1739861453886},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861453894,"name":"expect.toBe","stop":1739861453895},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861453899,"name":"expect.toBe","stop":1739861453894}],"attachments":[],"parameters":[],"start":1739861453311,"name":"beforeEach hook","stop":1739861453899},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861453908,"name":"browser.newContext","stop":1739861454051}],"attachments":[],"parameters":[],"start":1739861453904,"name":"fixture: context","stop":1739861454060},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861454063,"name":"browserContext.newPage","stop":1739861454174}],"attachments":[],"parameters":[],"start":1739861454060,"name":"fixture: page","stop":1739861454174}],"attachments":[],"parameters":[],"start":1739861453303,"name":"Before Hooks","stop":1739861454174},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861454191,"name":"page.goto(https://www.saucedemo.com)","stop":1739861454832},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861454840,"name":"locator.waitFor([data-test=\"username\"])","stop":1739861454898},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861454902,"name":"locator.fill([data-test=\"username\"])","stop":1739861454938},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861454942,"name":"locator.fill([data-test=\"password\"])","stop":1739861454979},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861454982,"name":"locator.click([data-test=\"login-button\"])","stop":1739861455148}],"attachments":[],"parameters":[],"start":1739861454177,"name":"Verifying loin to sauce demo","stop":1739861455149},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861455156,"name":"expect.toHaveText","stop":1739861455194}],"attachments":[],"parameters":[],"start":1739861455152,"name":"Main page header","stop":1739861455194},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861455205,"name":"page.screenshot","stop":1739861455473}],"attachments":[],"parameters":[],"start":1739861455201,"name":"page.screenshot","stop":1739861455466},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861455682,"name":"page.close","stop":1739861455690}],"attachments":[],"parameters":[],"start":1739861455474,"name":"afterEach hook","stop":1739861455690},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861455692,"name":"fixture: page","stop":1739861455687},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861455692,"name":"fixture: context","stop":1739861455693}],"attachments":[],"parameters":[],"start":1739861455195,"name":"After Hooks","stop":1739861455727},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"screenshot","source":"e29bcbad-0755-4857-bad0-220315f827ca-attachment.png","type":"image/png"}],"parameters":[],"start":1739861455903,"name":"screenshot","stop":1739861455903},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"screenshot","source":"03ff1a5e-febe-4659-8d0d-d654bbc232c5-attachment.png","type":"image/png"}],"parameters":[],"start":1739861455905,"name":"screenshot","stop":1739861455905},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"39aca89b-08ad-4bbd-8159-a6799dc34de7-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861455909,"name":"trace","stop":1739861455909}],"attachments":[{"name":"stdout","source":"3ac37328-d813-4e49-8027-2753f6b3bb8a-attachment.txt","type":"text/plain"}],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"ui.addToCart.spec.ts"},{"name":"titlePath","value":" > chromium > ui\\addToCart.spec.ts"},{"name":"tag","value":"regression"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"ui\\addToCart.spec.ts"}],"links":[],"start":1739861453280,"testCaseId":"9ba47862b673822ab6987e4d6c14df9f","fullName":"ui/addToCart.spec.ts:6:5","stop":1739861455512}
