--- allure-results/1fb2b3df-3bd6-47f7-b104-64a33355b393-attachment.txt
@@ -1,6 +1,0 @@
-New User: {
-  name: 'Kellie Gusikowski V',
-  job: 'Designer',
-  id: '292',
-  createdAt: '2025-02-18T06:50:46.132Z'
-}
