--- allure-results/3ac37328-d813-4e49-8027-2753f6b3bb8a-attachment.txt
@@ -1,5 +1,0 @@
-Set Page Worker Index: 0
-Environment set as: dev
-###### URL1=[object Object]
-###### URL2=https://www.saucedemo.com
-Sauce lab logged in
