--- allure-results/4333ca73-d573-45cf-9a13-cc0644d02e00-attachment.txt
@@ -1,2 +1,0 @@
-Set Page Worker Index: 0
-Environment set as: dev
