--- allure-results/48eb70ce-eb2f-48a6-b629-c2226d9fcc7b-attachment.txt
@@ -1,2 +1,0 @@
-###### URL1=[object Object]
-###### URL2=https://www.saucedemo.com
