--- allure-results/504a6a95-ba55-4d0d-8784-707df508d07f-result.json
@@ -1,1 +1,0 @@
-{"uuid":"504a6a95-ba55-4d0d-8784-707df508d07f","name":"Verify item added to cart successfully","historyId":"12cadc989ac8bbf2e22f1780f106fb01:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861472680,"name":"browser.newPage","stop":1739861473053},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861473058,"name":"expect.toBe","stop":1739861473059},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861473063,"name":"expect.toBe","stop":1739861473064}],"attachments":[],"parameters":[],"start":1739861472673,"name":"beforeEach hook","stop":1739861473064}],"attachments":[],"parameters":[],"start":1739861472672,"name":"Before Hooks","stop":1739861473065},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861473069,"name":"page.goto(https://www.saucedemo.com/inventory.html)","stop":1739861473809},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861473811,"name":"page.waitForTimeout","stop":1739861483829},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861483837,"name":"locator.click([data-test=\"add-to-cart-sauce-labs-backpack\"])","stop":1739861483931},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861483934,"name":"locator.click([data-test=\"shopping-cart-link\"])","stop":1739861484026},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861484033,"name":"locator.click([data-test=\"item-4-title-link\"])","stop":1739861484159}],"attachments":[],"parameters":[],"start":1739861483831,"name":"Adding to the cart","stop":1739861484159},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861484163,"name":"expect.toHaveText","stop":1739861484179}],"attachments":[],"parameters":[],"start":1739861484160,"name":"Verify Item Name In Cart","stop":1739861484179},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861484184,"name":"page.screenshot","stop":1739861484330},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861484443,"name":"page.close","stop":1739861484460}],"attachments":[],"parameters":[],"start":1739861484332,"name":"afterEach hook","stop":1739861484460}],"attachments":[],"parameters":[],"start":1739861484180,"name":"After Hooks","stop":1739861484461},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"screenshot","source":"9a127964-5252-4689-921e-99f53fb0570b-attachment.png","type":"image/png"}],"parameters":[],"start":1739861484637,"name":"screenshot","stop":1739861484637},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"d4097dd8-4a65-4c0f-95f5-aabbfdd5ff1c-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861484640,"name":"trace","stop":1739861484641}],"attachments":[{"name":"stdout","source":"4333ca73-d573-45cf-9a13-cc0644d02e00-attachment.txt","type":"text/plain"}],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"ui.loginSauceLogin.spec.ts"},{"name":"titlePath","value":" > chromium > ui\\loginSauceLogin.spec.ts > Single Family Activations tests"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"ui\\loginSauceLogin.spec.ts"},{"name":"subSuite","value":"Single Family Activations tests"}],"links":[],"start":1739861472721,"testCaseId":"12cadc989ac8bbf2e22f1780f106fb01","fullName":"ui/loginSauceLogin.spec.ts:19:9","stop":1739861484504}
