--- allure-results/6f27105b-153d-4dde-8211-5400e4384a9d-result.json
@@ -1,1 +1,0 @@
-{"uuid":"6f27105b-153d-4dde-8211-5400e4384a9d","name":"Test case to verify TTM integration with automation framework @SFAT-359","historyId":"ede43013141a2d10c2ae14fc64bdab18:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861469822,"name":"browserType.launch","stop":1739861470101}],"attachments":[],"parameters":[],"start":1739861469809,"name":"fixture: browser","stop":1739861470101},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861470109,"name":"browser.newPage","stop":1739861470716},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861470725,"name":"expect.toBe","stop":1739861470727},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861470733,"name":"expect.toBe","stop":1739861470734}],"attachments":[],"parameters":[],"start":1739861469780,"name":"beforeEach hook","stop":1739861470733}],"attachments":[],"parameters":[],"start":1739861469777,"name":"Before Hooks","stop":1739861470734},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861470756,"name":"page.goto(https://www.saucedemo.com)","stop":1739861471531},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861471546,"name":"locator.waitFor([data-test=\"username\"])","stop":1739861471630},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861471634,"name":"locator.fill([data-test=\"username\"])","stop":1739861471666},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861471668,"name":"locator.fill([data-test=\"password\"])","stop":1739861471698},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861471700,"name":"locator.click([data-test=\"login-button\"])","stop":1739861471841}],"attachments":[],"parameters":[],"start":1739861470740,"name":"Verifying loin to sauce demo","stop":1739861471842},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861471850,"name":"expect.toHaveText","stop":1739861471899}],"attachments":[],"parameters":[],"start":1739861471847,"name":"Main page header","stop":1739861471899},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861471906,"name":"locator.click([data-test=\"add-to-cart-sauce-labs-backpack\"])","stop":1739861471998},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861472003,"name":"locator.click([data-test=\"shopping-cart-link\"])","stop":1739861472050},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861472053,"name":"locator.click([data-test=\"item-4-title-link\"])","stop":1739861472137}],"attachments":[],"parameters":[],"start":1739861471901,"name":"Adding to the cart","stop":1739861472137},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861472143,"name":"expect.toHaveText","stop":1739861472166}],"attachments":[],"parameters":[],"start":1739861472139,"name":"Verify Item Name In Cart","stop":1739861472167},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861472172,"name":"page.screenshot","stop":1739861472306},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861472508,"name":"page.close","stop":1739861472514}],"attachments":[],"parameters":[],"start":1739861472310,"name":"afterEach hook","stop":1739861472514}],"attachments":[],"parameters":[],"start":1739861472168,"name":"After Hooks","stop":1739861472516},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"screenshot","source":"496aa2b2-bf6b-4f8f-a346-3638dccec538-attachment.png","type":"image/png"}],"parameters":[],"start":1739861472680,"name":"screenshot","stop":1739861472680},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"e75616c9-2399-4e70-8425-94a07ea8cdfe-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861472684,"name":"trace","stop":1739861472684}],"attachments":[{"name":"stdout","source":"c42731ba-9589-4e69-a997-978218947878-attachment.txt","type":"text/plain"}],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"ui.loginSauceLogin.spec.ts"},{"name":"titlePath","value":" > chromium > ui\\loginSauceLogin.spec.ts > Single Family Activations tests"},{"name":"tag","value":"SFAT-359"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"ui\\loginSauceLogin.spec.ts"},{"name":"subSuite","value":"Single Family Activations tests"}],"links":[],"start":1739861469776,"testCaseId":"ede43013141a2d10c2ae14fc64bdab18","fullName":"ui/loginSauceLogin.spec.ts:6:9","stop":1739861472207}
