--- allure-results/7ea6f3a9-ce18-4df4-90cd-83a20853641a-result.json
@@ -1,1 +1,0 @@
-{"uuid":"7ea6f3a9-ce18-4df4-90cd-83a20853641a","name":"Test4 add to cart","historyId":"9018e6a9ce1888f58914a62983b51bab:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861455931,"name":"browser.newPage","stop":1739861456167},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861456170,"name":"expect.toBe","stop":1739861456165},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861456171,"name":"expect.toBe","stop":1739861456172}],"attachments":[],"parameters":[],"start":1739861455911,"name":"beforeEach hook","stop":1739861456172},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861456174,"name":"browser.newContext","stop":1739861456311}],"attachments":[],"parameters":[],"start":1739861456172,"name":"fixture: context","stop":1739861456318},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861456322,"name":"browserContext.newPage","stop":1739861456467}],"attachments":[],"parameters":[],"start":1739861456319,"name":"fixture: page","stop":1739861456467}],"attachments":[],"parameters":[],"start":1739861455908,"name":"Before Hooks","stop":1739861456467},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861456470,"name":"page.goto(https://www.saucedemo.com/inventory.html)","stop":1739861457267},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861457270,"name":"page.waitForTimeout","stop":1739861467319},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861467325,"name":"locator.click([data-test=\"add-to-cart-sauce-labs-backpack\"])","stop":1739861467444},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861467450,"name":"locator.click([data-test=\"shopping-cart-link\"])","stop":1739861467595},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861467597,"name":"locator.click([data-test=\"item-4-title-link\"])","stop":1739861467697}],"attachments":[],"parameters":[],"start":1739861467320,"name":"Adding to the cart","stop":1739861467697},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861467703,"name":"expect.toHaveText","stop":1739861467733}],"attachments":[],"parameters":[],"start":1739861467699,"name":"Verify Item Name In Cart","stop":1739861467733},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861467742,"name":"page.screenshot","stop":1739861468063}],"attachments":[],"parameters":[],"start":1739861467739,"name":"page.screenshot","stop":1739861468064},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861468163,"name":"page.close","stop":1739861468171}],"attachments":[],"parameters":[],"start":1739861468066,"name":"afterEach hook","stop":1739861468171},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861468172,"name":"fixture: page","stop":1739861468168},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861468173,"name":"fixture: context","stop":1739861468168}],"attachments":[],"parameters":[],"start":1739861467734,"name":"After Hooks","stop":1739861468200},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"screenshot","source":"331bbca9-2a38-41df-8b5c-08e88721ffee-attachment.png","type":"image/png"}],"parameters":[],"start":1739861468343,"name":"screenshot","stop":1739861468343},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"screenshot","source":"2d9471fc-8fe0-46a1-9dd2-97c6a440635a-attachment.png","type":"image/png"}],"parameters":[],"start":1739861468345,"name":"screenshot","stop":1739861468345},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"7e3fcbb3-d8ff-4613-9781-afe2df2073ba-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861468346,"name":"trace","stop":1739861468346}],"attachments":[{"name":"stdout","source":"e83ea30c-37c6-435e-9156-acf62168fbc5-attachment.txt","type":"text/plain"}],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"ui.addToCart.spec.ts"},{"name":"titlePath","value":" > chromium > ui\\addToCart.spec.ts"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"ui\\addToCart.spec.ts"}],"links":[],"start":1739861455926,"testCaseId":"9018e6a9ce1888f58914a62983b51bab","fullName":"ui/addToCart.spec.ts:12:5","stop":1739861468205}
