--- allure-results/9f47f0c1-d56a-4557-87e9-53921b5574a4-result.json
@@ -1,1 +1,0 @@
-{"uuid":"9f47f0c1-d56a-4557-87e9-53921b5574a4","name":"Test case to verify TTM integration with automation framework @SFAT-359","historyId":"ede43013141a2d10c2ae14fc64bdab18:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"failed","statusDetails":{"message":"Error: page.goto: Target page, context or browser has been closed","trace":"Error: page.goto: Target page, context or browser has been closed\n    at gotoURL (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\utils\\action-utils.ts:41:26)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\pages\\Login\\login-page.ts:17:22\n    at Object.loginToSauceDemo (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\pages\\Login\\login-page.ts:14:16)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\tests\\e2e\\ui\\loginSauceLogin.spec.ts:12:25"},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861468373,"name":"Before Hooks","stop":1739861468376},{"status":"failed","statusDetails":{"message":"Error: page.goto: Target page, context or browser has been closed","trace":"Error: page.goto: Target page, context or browser has been closed\n    at gotoURL (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\utils\\action-utils.ts:41:26)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\pages\\Login\\login-page.ts:17:22\n    at Object.loginToSauceDemo (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\pages\\Login\\login-page.ts:14:16)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\tests\\e2e\\ui\\loginSauceLogin.spec.ts:12:25"},"stage":"finished","steps":[{"status":"failed","statusDetails":{"message":"Error: page.goto: Target page, context or browser has been closed","trace":"Error: page.goto: Target page, context or browser has been closed\n    at gotoURL (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\utils\\action-utils.ts:41:26)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\pages\\Login\\login-page.ts:17:22\n    at Object.loginToSauceDemo (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\pages\\Login\\login-page.ts:14:16)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\tests\\e2e\\ui\\loginSauceLogin.spec.ts:12:25"},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861468380,"name":"page.goto(https://www.saucedemo.com)","stop":1739861468382}],"attachments":[],"parameters":[],"start":1739861468377,"name":"Verifying loin to sauce demo","stop":1739861468383},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861468384,"name":"After Hooks","stop":1739861468386},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861468387,"name":"fixture: browser","stop":1739861468484}],"attachments":[],"parameters":[],"start":1739861468386,"name":"Worker Cleanup","stop":1739861468485},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"f6c05276-e8d5-4e83-b9fe-bf06061488aa-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861468527,"name":"trace","stop":1739861468527}],"attachments":[{"name":"stdout","source":"48eb70ce-eb2f-48a6-b629-c2226d9fcc7b-attachment.txt","type":"text/plain"}],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"ui.loginSauceLogin.spec.ts"},{"name":"titlePath","value":" > chromium > ui\\loginSauceLogin.spec.ts > Single Family Activations tests"},{"name":"tag","value":"SFAT-359"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"ui\\loginSauceLogin.spec.ts"},{"name":"subSuite","value":"Single Family Activations tests"}],"links":[],"start":1739861468380,"testCaseId":"ede43013141a2d10c2ae14fc64bdab18","fullName":"ui/loginSauceLogin.spec.ts:6:9","stop":1739861468390}
