--- allure-results/c42731ba-9589-4e69-a997-978218947878-attachment.txt
@@ -1,5 +1,0 @@
-Set Page Worker Index: 0
-Environment set as: dev
-###### URL1=[object Object]
-###### URL2=https://www.saucedemo.com
-Sauce lab logged in
