--- allure-results/c54762ea-2f4a-4213-b4a7-82c61cba84a3-result.json
@@ -1,1 +1,0 @@
-{"uuid":"c54762ea-2f4a-4213-b4a7-82c61cba84a3","name":"Create a new user and verify the response","historyId":"b3c458ab846d3b1ae143c91224102915:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861445707,"name":"Before Hooks","stop":1739861445738},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861445752,"name":"apiRequestContext.post(https://reqres.in/api/users)","stop":1739861446216},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446221,"name":"apiResponse.json","stop":1739861446223},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446230,"name":"expect.toBeDefined","stop":1739861446232},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446234,"name":"expect.toBeDefined","stop":1739861446231},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446237,"name":"expect.toBe","stop":1739861446238},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446241,"name":"expect.toBe","stop":1739861446238},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446249,"name":"After Hooks","stop":1739861446330},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"e10b9911-efd8-468b-b83c-1b0521dad9fa-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861446380,"name":"trace","stop":1739861446380}],"attachments":[{"name":"stdout","source":"1fb2b3df-3bd6-47f7-b104-64a33355b393-attachment.txt","type":"text/plain"}],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"api.users.spec.ts"},{"name":"titlePath","value":" > chromium > api\\users.spec.ts"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"api\\users.spec.ts"}],"links":[],"start":1739861445700,"testCaseId":"b3c458ab846d3b1ae143c91224102915","fullName":"api/users.spec.ts:11:5","stop":1739861446225}
