--- allure-results/ca32cdec-3a16-4c56-a90f-46ce27f6e01d-result.json
@@ -1,1 +1,0 @@
-{"uuid":"ca32cdec-3a16-4c56-a90f-46ce27f6e01d","name":"Get user details by ID and verify the response","historyId":"db691e99913cd1b92ce4aa29e66c89f1:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"failed","statusDetails":{"message":"Error: Expected status 200, but got 404: {}","trace":"Error: Expected status 200, but got 404: {}\n    at ApiClient.request (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\api\\apiClient.ts:59:15)\n    at UserService.getUserDetails (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\api\\services\\usersService.ts:34:22)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\tests\\e2e\\api\\users.spec.ts:37:25"},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446368,"name":"Before Hooks","stop":1739861446377},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446381,"name":"apiRequestContext.post(https://reqres.in/api/users)","stop":1739861446652},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446658,"name":"apiResponse.json","stop":1739861446662},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446666,"name":"expect.toBeDefined","stop":1739861446663},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446670,"name":"expect.toBeDefined","stop":1739861446667},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446674,"name":"expect.toBe","stop":1739861446671},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446678,"name":"expect.toBe","stop":1739861446675},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446684,"name":"apiRequestContext.get(https://reqres.in/api/users/875)","stop":1739861446991},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861446999,"name":"apiResponse.text","stop":1739861447003},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861447006,"name":"After Hooks","stop":1739861447060},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861447062,"name":"Worker Cleanup","stop":1739861447063},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"a40e6789-233b-45fb-9838-b56fd4bae62b-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861447106,"name":"trace","stop":1739861447106}],"attachments":[],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"api.users.spec.ts"},{"name":"titlePath","value":" > chromium > api\\users.spec.ts"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"api\\users.spec.ts"}],"links":[],"start":1739861446393,"testCaseId":"db691e99913cd1b92ce4aa29e66c89f1","fullName":"api/users.spec.ts:24:5","stop":1739861447025}
