--- allure-results/dcd54297-af4a-498b-af04-4a67caf0a721-result.json
@@ -1,1 +1,0 @@
-{"uuid":"dcd54297-af4a-498b-af04-4a67caf0a721","name":"Get user details by ID and verify the response","historyId":"db691e99913cd1b92ce4aa29e66c89f1:5bd835b0d6b1d4ada3b9f0db936e82c8","status":"failed","statusDetails":{"message":"Error: Expected status 200, but got 404: {}","trace":"Error: Expected status 200, but got 404: {}\n    at ApiClient.request (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\api\\apiClient.ts:59:15)\n    at UserService.getUserDetails (D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\src\\api\\services\\usersService.ts:34:22)\n    at D:\\Conservice\\workspace\\github\\conservice\\conservice-qa-automation-poc\\tests\\e2e\\api\\users.spec.ts:37:25"},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451150,"name":"Before Hooks","stop":1739861451169},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451179,"name":"apiRequestContext.post(https://reqres.in/api/users)","stop":1739861451603},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451607,"name":"apiResponse.json","stop":1739861451609},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451616,"name":"expect.toBeDefined","stop":1739861451617},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451621,"name":"expect.toBeDefined","stop":1739861451623},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451624,"name":"expect.toBe","stop":1739861451625},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451627,"name":"expect.toBe","stop":1739861451628},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861451632,"name":"apiRequestContext.get(https://reqres.in/api/users/869)","stop":1739861452009},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861452015,"name":"apiResponse.text","stop":1739861452026},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861452030,"name":"After Hooks","stop":1739861452120},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1739861452121,"name":"Worker Cleanup","stop":1739861452123},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"f07e79e2-5025-42a9-9f42-b8e8781317cb-attachment.zip","type":"application/zip"}],"parameters":[],"start":1739861452168,"name":"trace","stop":1739861452168}],"attachments":[],"parameters":[{"name":"Project","value":"chromium"}],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"api.users.spec.ts"},{"name":"titlePath","value":" > chromium > api\\users.spec.ts"},{"name":"host","value":"DIC-LAP-147"},{"name":"thread","value":"pid-14664-worker-0"},{"name":"parentSuite","value":"chromium"},{"name":"suite","value":"api\\users.spec.ts"}],"links":[],"start":1739861451148,"testCaseId":"db691e99913cd1b92ce4aa29e66c89f1","fullName":"api/users.spec.ts:24:5","stop":1739861452018}
