--- allure-results/e83ea30c-37c6-435e-9156-acf62168fbc5-attachment.txt
@@ -1,3 +1,0 @@
-Set Page Worker Index: 0
-Environment set as: dev
-Test2 still logged in using storage state
