+++ config/ui/globalSetup.ts
