+++ tests/api/users.spec.ts
@@ -1,11 +1,11 @@
 import { test, expect } from '@playwright/test';
-import { ApiClient } from '../../../src/api/apiClient';
-import { UserService } from '../../../src/api/services/usersService';
-import { generateRandomUser } from '../../../tests/testdata/user-data';
+import { ApiClient } from '../../src/api/apiClient';
+import { UserService } from '../../src/api/services/usersService';
+import { generateRandomUser } from '../../src/api/api-utils/user-data';
 
 
 const client = new ApiClient('https://reqres.in');
-client.authenticateWithCredentials("/test","test","test");
+client.authenticateWithCredentials("/test", "test", "test");
 const userService = new UserService(client);
 
 test('Create a new user and verify the response', async () => {
@@ -28,16 +28,16 @@ test('Get user details by ID and verify the response', async () => {
   };
 
   const newUser = await userService.createUser(userData, 201);
-    expect(newUser.id).toBeDefined();
-    expect(newUser.createdAt).toBeDefined();
-    expect(newUser.name).toBe(userData.name);
-    expect(newUser.job).toBe(userData.job);
-  
-    const userId = newUser.id.toString();
-    const userDetails = await userService.getUserDetails(userId, 200);
-    expect(userDetails.id).toBeDefined();
-    expect(userDetails.name).toBe(userData.name);
-    expect(userDetails.job).toBe(userData.job);
+  expect(newUser.id).toBeDefined();
+  expect(newUser.createdAt).toBeDefined();
+  expect(newUser.name).toBe(userData.name);
+  expect(newUser.job).toBe(userData.job);
+
+  const userId = newUser.id.toString();
+  const userDetails = await userService.getUserDetails(userId, 200);
+  expect(userDetails.id).toBeDefined();
+  expect(userDetails.name).toBe(userData.name);
+  expect(userDetails.job).toBe(userData.job);
 
   console.log('User Details:', userDetails);
 });
